make WITH <uri> DELETE ... INSERT ... WHERE .. SPARQL 1.1 update (draft) work
Because the grammar is ambiguous in several places and unstable, in
this grammar WITH <uri> is required, not optional as the grammar
makes it.

Shift/reduce conflicts goes from 5 to 13 probably because of
insert/delete ambiguity.

(ConstructTriples): Added; was part of ConstructTriplesOpt

(GraphTemplate): Added.

(ModifyTemplate, ModifyTemplateList): To form list of graphs to
update.  This data is thrown away now, not stored.

(UpdateQuery): Added but pretending to be INSERT verb.
Only supports  WITH <uri> DELETE { ... } INSERT { ... } WHERE ...
form with no optional WITH; otherwise this is even more ambiguous.