Re-add SPARQL 1.1 WITH .. DELETE/INSERT
From 5 shift/reduce conflicts to 31 shift/reduce conflicts - seems to
be choices between two optional sequences of triples nearby.

Add UpdateQuery option that (for now) marks it as an INSERT.

(DeleteQuery): Switch to list of modify templates - which are
discarded.

(GraphTemplate): Added
(ModifyTemplate, ModifyTemplateList): Added.

(InsertQuery): Switch to list of modify templates - which are
discarded.

(UpdateQuery): Added for WITH alternatives.