Add -lraptor to internal linking of utility and example programs.
Added -lraptor to linking roqet, srxread and srxwrite so that on
systems where indirect library symbols (here in libraptor) are not
automatically found by looking up the base library (here librasqal),
the programs work.  This should not affect user programs that link
with rasqal since they use
  pkg-config rasqal --libs
or
  rasqal-config --libs
both of which add -lraptor.

See also
   Fedora bug 564885 for rasqal
   https://bugzilla.redhat.com/show_bug.cgi?id=564885
and
   http://fedoraproject.org/wiki/Features/ChangeInImplicitDSOLinking
   https://fedoraproject.org/wiki/UnderstandingDSOLinkChange

Fixes Issue#0000347
http://bugs.librdf.org/mantis/view.php?id=347