Handle LOAD ... list of DocUris... (INTO GraphUri)
(IriRefList): Added to build a raptor_sequence of raptor_uri

(LoadQuery): Use IriRefList to iterate over doc uris and make an
update operation for each of them.  This may not correspond to the
actual sparql update semantics, to be determined. If the list-load is
one operation, it will need a rasqal_update_operation API change