Add support for BINDINGS variables and lists of rows
Extended parser token to contain a rasqal_row

(BindingsRowList, BindingsRowListOpt): Added as
sequences of rasqal_row

(BindingValueList): Added as sequence of rasqal_literals

(BindingsRow): Added as a rasqal_row

(VarList): Added as a sequence of variables

(BindingsClause): Rewritten to follow SPARQL 1.1 grammar and
construct a sequence of variables and sequence of rows.  All are
discarded rather than stored for now.