Code style change and cleanup for alloc/free macros
Code style:
  1. var = RASQAL_CALLOC(type, nmem, size)
  Prefering:
    var = RASQAL_CALLOC(type, 1, sizeof(*var))
  when nmem = 1

  2. var = RASQAL_CALLOC(type, size)

  3. RASQAL_FREE(type, var)

The consequence here is allocs that mostly fit into 1 line without so
much boilerplate and duplication of types.

The RASQAL_MALLOC and RASQAL_CALLOC now do the cast to the return type.

RASQAL_FREE takes the object type too but always casts arg to void
This certainly contains many wrong types to the arg but might be
used later in some kind of smart type-aware debugging allocator.